+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+- divided buttons on bars
+- added hill shade
 - changed to use Tile server url
 - added click on monitoring location to add show location info on information bar
 - added hru hover that displays name in the information bar